Time: 37 ms (36.99%), Space: 21.5 MB (15.29%) - LeetHub

+++ 0875-koko-eating-bananas/0875-koko-eating-bananas.cpp
@@ -5,7 +5,7 @@ public:
         int hi = *std::max_element(piles.begin(), piles.end());
         while (lo <= hi) {
             int k = (hi - lo) / 2 + lo;
-            int kSum = 0;
+            long kSum = 0;
             for (int pile : piles) {
                 kSum += std::ceil(static_cast<double>(pile) / k);
             }
